Update LICENSE to reflect non-open-source EULA
Railway is not open source.  Use of the Railway wallet software explicitly requires agreement to the terms of use (EULA), which is a license incompatible with the concept of open source.

Either the AGPL applies, or the Railway EULA (presented as a requirement in the UI) applies.  It cannot be both.

As the software requires explicit confirmation of the terms of the EULA, the software is not open source.  Therefore, an open source LICENSE.md file is inaccurate and should be corrected.  This PR updates the LICENSE terms to be those presented within the application - the proprietary end user *LICENSE* agreement.